cleanup
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@2972 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-app/src/main/java/org/sonatype/nexus/tasks/EvictUnusedProxiedItemsTask.java
@@ -51,19 +51,19 @@ public class EvictUnusedProxiedItemsTask
         if ( getRepositoryGroupId() != null )
         {
             return getNexus().evictRepositoryGroupUnusedProxiedItems(
-                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * A_DAY ),
+                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * A_DAY ),
                 getRepositoryGroupId() );
         }
         else if ( getRepositoryId() != null )
         {
             return getNexus().evictRepositoryUnusedProxiedItems(
-                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * A_DAY ),
+                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * A_DAY ),
                 getRepositoryId() );
         }
         else
         {
             return getNexus().evictAllUnusedProxiedItems(
-                System.currentTimeMillis() - ( ( (long) getEvictOlderCacheItemsThen() ) * A_DAY ) );
+                System.currentTimeMillis() - ( getEvictOlderCacheItemsThen() * A_DAY ) );
         }
     }
